+++ Tools/CppToCsharpConverter/Converter.Golos/Converters/BaseConverter.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Converter.Core;
+using Converter.Core.Helpers;
 using Converter.Core.Models;
 using Newtonsoft.Json.Linq;
 
@@ -120,7 +121,7 @@ namespace Converter.Golos.Converters
                 return;
 
             if (!string.IsNullOrEmpty(parsedClass.CppName))
-                parsedClass.Name = _cashParser.ToTitleCase(parsedClass.CppName);
+                parsedClass.Name = parsedClass.CppName.ToTitleCase();
             if (!string.IsNullOrEmpty(parsedClass.CppInherit))
                 parsedClass.Inherit = new List<ParsedType> { GetKnownTypeOrDefault(parsedClass.CppInherit) };
 
@@ -194,7 +195,7 @@ namespace Converter.Golos.Converters
             if (!NotNameChar.IsMatch(type))
             {
                 AddTypeToTask(type);
-                return new ParsedType { CppName = type, Name = _cashParser.ToTitleCase(type), IsOptional = isOptional };
+                return new ParsedType { CppName = type, Name = type.ToTitleCase(), IsOptional = isOptional };
             }
 
             return new ParsedType { CppName = type, Name = "object", IsOptional = isOptional };
@@ -204,7 +205,7 @@ namespace Converter.Golos.Converters
         {
             if (!Founded.ContainsKey(type))
             {
-                Founded.Add(type, _cashParser.ToTitleCase(type));
+                Founded.Add(type, type.ToTitleCase());
                 if (!UnknownTypes.Any(i => i.SearchLine.Equals(type) && string.IsNullOrEmpty(i.SearchDir)))
                 {
                     var task = new SearchTask
@@ -293,7 +294,7 @@ namespace Converter.Golos.Converters
                 var param = new ParsedParams()
                 {
                     CppName = name,
-                    Name = _cashParser.ToTitleCase(name, false),
+                    Name = name.ToTitleCase(false),
                     Default = defaultValue,
                     CppType = typeMatches[i].Value,
                     Type = GetKnownTypeOrDefault(typeMatches[i].Value)
@@ -329,7 +330,7 @@ namespace Converter.Golos.Converters
             var rez = new ParsedParams()
             {
                 CppName = name,
-                Name = _cashParser.ToTitleCase(name, false),
+                Name = name.ToTitleCase(false),
                 Default = defaultValue,
                 CppType = typeMatche.Value,
                 Type = GetKnownTypeOrDefault(typeMatche.Value)
