+++ Tools/CppToCsharpConverter/Converter.Golos/Converters/StructConverter.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using Converter.Core.Helpers;
 using Converter.Core.Models;
 
 namespace Converter.Golos.Converters
@@ -93,7 +94,7 @@ namespace Converter.Golos.Converters
             var field = new PreParsedElement
             {
                 Type = GetKnownTypeOrDefault(cppType, templateName),
-                Name = _cashParser.ToTitleCase(name),
+                Name = name.ToTitleCase(),
                 CppName = name,
                 Comment = coment,
                 MainComment = preParsedElement.MainComment
