+++ Tools/CppToCsharpConverter/Converter.Steem/Converters/StructConverter.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using Converter.Steem.Converters;
 using Converter.Core;
+using Converter.Core.Helpers;
 using Converter.Core.Models;
 
 namespace Converter.Steem
@@ -94,7 +95,7 @@ namespace Converter.Steem
             var field = new PreParsedElement
             {
                 Type = GetKnownTypeOrDefault(cppType, templateName),
-                Name = _cashParser.ToTitleCase(name),
+                Name = name.ToTitleCase(),
                 CppName = name,
                 Comment = coment,
                 MainComment = preParsedElement.MainComment
